requirements.txt
bar-chart-race==0.1.0
certifi==2022.6.15
chardet==4.0.0
cycler==0.11.0
ffmpeg-python==0.2.0
fonttools==4.34.4
FormEncode==2.0.1
future==0.18.2
idna==2.10
kiwisolver==1.4.3
matplotlib==3.5.2
mysqlclient==2.1.1
numpy==1.23.0
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
PyDispatcher==2.0.5
pyparsing==2.4.7
python-crontab==2.6.0
python-dateutil==2.8.1
python-dotenv==0.20.0
python-twitter==3.5
pytz==2022.1
requests==2.25.1
requests-oauthlib==1.3.1
six==1.16.0
SQLObject==3.9.1
tweet-preprocessor==0.6.0
urllib3==1.26.10
wordcloud==1.8.2.2